chore: various improvements
