+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "nyahaha-bot",
-  "version": "0.0.12",
+  "version": "0.0.13",
   "description": "The code base for a personal bot.",
   "main": "dist/index.js",
   "scripts": {
@@ -48,9 +48,9 @@
     "crypto-js": "^4.1.1",
     "dayjs": "^1.11.10",
     "express": "^4.18.2",
-    "firebase-admin": "^11.10.1",
+    "firebase-admin": "^12.1.0",
     "join-images": "^1.1.5",
-    "node-telegram-bot-api": "^0.64.0",
+    "node-telegram-bot-api": "^0.63.0",
     "opencc-js": "^1.0.5",
     "prompts": "^2.4.2",
     "sharp": "^0.32.6",
